test-runner: Disable address sanitization on .test_section
Clang inserts red-zones around allocations in user defined section, which break
the common assumption that objects are densely packed. The test runner uses this
trick to construct an array of test objects, which causes asan violations when
iterating over the array.

This can be disabled by adding __attribute__((no_sanitize("address"))) to each
object in the section.

See https://github.com/google/sanitizers/issues/1028 for more discussion of this issue.

Signed-off-by: Fergus Dall <sidereal@google.com>