protocol: add wl_pointer.warp event
The original wl_pointer.button event does not contain the position information.
In some scenarios, the client cannot obtain the correct position information.
For example, when the mouse focus is on a button and the mouse does not move
but the window size changes, the client cannot update the position information
through the motion event, resulting in incorrect mouse position.

Therefore, This event should be sent when the local coordinates of a pointer
change as a result of surface being moved/resized/fullscreened.

Signed-off-by: Lu YaNing <luyaning@uniontech.com>