wayland-server: Add wl_client_add_event_dispatch_listener
The patch adds a listener called after an event has been dispatched
by the server.

Wayland servers that incorporate the Wayland display into their own
event loops may dispatch multiple Wayland events to clients as they
work through their task queues.

For e.g. if system display state changes the Wayland server may
emit several output / surface Wayland events to connected clients.

Currently an explicit flush is performd immediately after each
individual Wayland event is dispatched to make sure these are
propagated to clients in a deterministic / reliable manner.

This results in hard to maintain / easy to break code. Contributors
must make sure to explicitly flush when dispatching Wayland events
to clients - or otherwise wait for some other server event to flush
the client buffer later on. In the latter case behavior is not
deterministic and makes testing the server code much harder.

The client event_dispatch_listener added in this patch allows the
server to track if the processing of a server event has resulted in
the dispatch of one or more Wayland events to clients - and ensure
the buffer is appropriately flushed.

This eliminates the need to flush in many different areas of the
codebase, making the server easier to maintain and test.

Signed-off-by: Thomas Lukaszewicz <thomaslukaszewicz@gmail.com>