client/server: add delete_id request as a fake sync
Add a delete_id request so that the server can reap
zombies once the client acks the deletion.  As a
result, the wl_map zombie_list would only be used
for clients that do not have this change.

To keep compatibility, the delete_id request is
dressed as a sync request, but carries a server
object ID instead.

Update tests to compensate for additional handshake
that sets up delete_id requests.

Signed-off-by: Jonathan Leivent <jleivent@comcast.net>