client: print a warning for undispatched queues when debugging
Forgetting to dispatch a queue can be very hard to debug, so let's try
to make that a little easier by printing a warning if any queue has
undispatched events for more than 3 seconds.

We can only check this when we're in libwayland for some reason, so let's
check when marshalling a request or when dispatching events on any queue.

Signed-off-by: Derek Foreman <derek.foreman@collabora.com>