protocol: frame requests run on a schedule
There are applications that expect frame requests to be throttled even
if the surface contents are unchanged. In situations where the
compositor wants the application to make progress, these applications
also expect frame requests to be signaled frequently.

Theses two requirements do not always correspond to any external
schedule imposed on the compositor. For example, consider a compositor
running a single full-screen surface on an output device with an
infinite VRR range. In such a situtation, the compositor has to create
an artificial schedule on which to dispatch frame requests.

Signed-off-by: Julian Orth <ju.orth@gmail.com>