Document that wl_shm.format is not authoritative
Compositors are allowed to announce formats that are not in the
wl_shm.format enum, and clients are allowed to use these formats.
Therefore, the enum="format" and enum="wl_shm.format" attributes
in the XML are unhelpful at best and misleading at worst.  Clients
and compositors should reference drm_fourcc.h instead.

Signed-off-by: Demi Marie Obenour <demi@invisiblethingslab.com>